chg: isOS install
